Initial splitting out of GenShader Utilities.
